backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-jose[cryptography]
passlib[argon2]
python-multipart
emails
pydantic-settings
argon2-cffi
